removed ModelAdapter and renamed module to 'executor.py'
